Move dq.rand namespace to dq.rand_*
