+++ docs/python_api/index.md
@@ -165,10 +165,17 @@ The **dynamiqs** Python API features two main types of functions: solvers of dif
     options:
         table: true
         members:
-        - rand_real
-        - rand_complex
         - snap_gate
         - cd_gate
+
+### Random arrays
+
+::: dynamiqs.utils.random
+    options:
+        table: true
+        members:
+        - rand_real
+        - rand_complex
 
 ### Plotting
