+++ dynamiqs/utils/__init__.py
@@ -1,6 +1,7 @@
 from .array_types import *
 from .operators import *
 from .optimal_control import *
+from .random import *
 from .states import *
 from .utils import *
 from .vectorization import *
