+++ dynamiqs/utils/random.py
@@ -1,0 +1,123 @@
+from __future__ import annotations
+
+import jax
+from jax import Array
+from jaxtyping import PRNGKeyArray
+
+from .utils import dag, unit
+
+__all__ = ['rand_real', 'rand_complex', 'rand_herm', 'rand_psd', 'rand_dm', 'rand_ket']
+
+
+def rand_real(
+    key: PRNGKeyArray,
+    shape: int | tuple[int, ...],
+    *,
+    min: float = 0.0,
+    max: float = 1.0,
+) -> Array:
+    r"""Returns an array filled with uniformly distributed random real numbers.
+
+    Each element of the returned array is sampled uniformly in
+    $[\text{min}, \text{max})$.
+
+    Args:
+        key: A PRNG key used as the random key.
+        shape _(int or tuple of ints)_: Shape of the returned array.
+        min: Minimum (inclusive) value.
+        max: Maximum (exclusive) value.
+
+    Returns:
+        _(array of shape (*shape))_ Array filled with random real numbers.
+
+    Examples:
+        >>> key = jax.random.PRNGKey(42)
+        >>> dq.rand_real(key, (2, 5), max=5.0)
+        Array([[3.22 , 1.613, 0.967, 4.432, 4.21 ],
+               [0.96 , 1.726, 1.262, 3.16 , 3.274]], dtype=float32)
+    """
+    shape = (shape,) if isinstance(shape, int) else shape
+    # sample uniformly in [min, max)
+    return jax.random.uniform(key, shape=shape, minval=min, maxval=max)
+
+
+def rand_complex(
+    key: PRNGKeyArray, shape: int | tuple[int, ...], *, rmax: float = 1.0
+) -> Array:
+    r"""Returns an array filled with random complex numbers uniformly distributed in
+    the complex plane.
+
+    Each element of the returned array is sampled uniformly in the disk of radius
+    $\text{rmax}$.
+
+    Notes-: Uniform sampling in the complex plane
+        Here are three common options to generate random complex numbers,
+        `dq.rand_complex()` returns the last one:
+
+        ```python
+        _, (ax0, ax1, ax2) = dq.gridplot(3, sharex=True, sharey=True)
+        ax0.set(xlim=(-1.1, 1.1), ylim=(-1.1, 1.1))
+
+        n = 10_000
+
+        # option 1: uniformly distributed real and imaginary part
+        x = np.random.rand(n) * 2 - 1 + 1j * (np.random.rand(n) * 2 - 1)
+        ax0.scatter(x.real, x.imag, s=1.0)
+
+        # option 2: uniformly distributed magnitude and phase
+        x = np.random.rand(n) * np.exp(1j * 2 * np.pi * np.random.rand(n))
+        ax1.scatter(x.real, x.imag, s=1.0)
+
+        # option 3: uniformly distributed in a disk (in dynamiqs)
+        key = jax.random.PRNGKey(42)
+        x = dq.rand_complex(key, n)
+        ax2.scatter(x.real, x.imag, s=1.0)
+        renderfig('rand_complex')
+        ```
+
+        ![rand_complex](/figs-code/rand_complex.png){.fig}
+
+    Args:
+        key: A PRNG key used as the random key.
+        shape _(int or tuple of ints)_: Shape of the returned array.
+        rmax: Maximum magnitude.
+
+    Returns:
+        _(array of shape (*shape))_ Array filled with random complex numbers.
+
+    Examples:
+        >>> key = jax.random.PRNGKey(42)
+        >>> dq.rand_complex(key, (2, 3), rmax=5.0)
+        Array([[ 1.341+4.17j ,  3.978-0.979j, -2.592-0.946j],
+               [-4.428+1.744j, -0.53 +1.668j,  2.582+0.65j ]], dtype=complex64)
+    """
+    shape = (shape,) if isinstance(shape, int) else shape
+    # sample uniformly in the unit L2 ball and scale
+    x = rmax * jax.random.ball(key, 2, shape=shape)
+    return x[..., 0] + 1j * x[..., 1]
+
+
+def rand_herm(key: PRNGKeyArray, shape: tuple[int, ...]) -> Array:
+    # hermitian
+    assert len(shape) >= 2 and shape[-1] == shape[-2]
+    x = rand_complex(key, shape)
+    return 0.5 * (x + dag(x))
+
+
+def rand_psd(key: PRNGKeyArray, shape: tuple[int, ...]) -> Array:
+    # positive semi-definite
+    assert len(shape) >= 2 and shape[-1] == shape[-2]
+    x = rand_complex(key, shape)
+    return x @ dag(x)
+
+
+def rand_dm(key: PRNGKeyArray, shape: tuple[int, ...]) -> Array:
+    assert len(shape) >= 2 and shape[-1] == shape[-2]
+    x = rand_psd(key, shape)
+    return unit(x)
+
+
+def rand_ket(key: PRNGKeyArray, shape: tuple[int, ...]) -> Array:
+    assert len(shape) >= 2 and shape[-1] == 1
+    x = rand_complex(key, shape)
+    return unit(x)
